bug fixed

+++ layers/maskNet.py
@@ -66,7 +66,7 @@ class PyramidROIAlign(keras.layers.Layer):
         # Equation 2 of CenterMask(https://arxiv.org/abs/1911.06667)
         roi_level = log2_graph(h*w)
         roi_level = tf.minimum(5, tf.maximum(
-            3, tf.math.ceil(5 + roi_level)))
+            3, tf.cast(tf.math.ceil(5.0 + roi_level), dtype=tf.int32)))
         roi_level = tf.squeeze(roi_level, 2)
 
         # Loop through levels and apply ROI pooling to each. P3 to P6.
